requirements-messy.txt
adminui==0.3.7
aenum==2.2.4
asgiref==3.2.10
auth==0.5.3
blinker==1.4
blist==1.3.6
cassandra-driver==3.4.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cqlengine==0.21.0
Django==3.1.1
django-cassandra-engine==1.6.1
dnspython==1.16.0
environs==8.0.0
eventlet==0.27.0
falcon==2.0.0
Flask==1.1.2
Flask-Admin==1.5.6
Flask-Cassandra==0.14
Flask-CQLAlchemy==1.2.0
Flask-WTF==0.14.3
futures==3.1.1
geomet==0.2.1.post1
greenlet==0.4.16
gremlinpython==3.4.6
gunicorn==20.0.4
idna==2.10
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.8.0
mongoengine==0.20.0
monotonic==1.5
PyJWT==1.7.1
pymongo==3.11.0
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
tornado==5.1.1
urllib3==1.25.10
Werkzeug==0.16.0
WTForms==2.3.3
